Implement management of multiple price and stock combinations
