+++ components/admin/index.ts
@@ -1,3 +1,11 @@
 export * from './AdminNavbar';
 export * from './SummaryTile';
 export * from './FormProduct';
+export * from './ProductInfo';
+export * from './ProductConfiguration';
+export * from './BrandModelInput';
+export * from './TagListDisplay';
+export * from './ImageManagement';
+export * from './TypeSelection';
+export * from './FeatureSelection';
+export * from './PriceAndStockVariations';
