+++ constants/productConstants.ts
@@ -11,3 +11,8 @@ export const SHOP_CONSTANTS = {
   validCapacity: ['64GB', '128GB', '256GB', '512GB', '1TB', '2TB'],
   validRam: ['4GB', '8GB', '12GB', '16GB', '32GB', '64GB'],
 };
+
+export const FEATURE_OPTIONS = [
+  { name: 'Capacidad', type: 'capacity' },
+  { name: 'Ram', type: 'ram' }
+];
